Factor out DropRoleTask from TransportDropRoleAction
